+++ src/validators.js
@@ -398,7 +398,7 @@ extend(angularValidator, {
           $invalidWidgets.markValid(element);
         }
         element.data($$validate)();
-        scope.$root.$eval();
+        scope.$updateView();
       });
     } else if (inputState.inFlight) {
       // request in flight, mark widget invalid, but don't show it to user