+++ test/servicesSpec.js
@@ -382,7 +382,7 @@ describe("service", function(){
     it('should call eval even if an exception is thrown in callback', function() {
       var eval = this.spyOn(scope, '$eval').andCallThrough();
 
-      $defer(function() {throw "Test Error"});
+      $defer(function() {throw "Test Error";});
       expect(eval).wasNotCalled();
 
       $browser.defer.flush();
@@ -582,7 +582,7 @@ describe("service", function(){
 
         $browser.defer.flush();
         expect(eval).wasCalled();
-      })
+      });
     });
 
   });
@@ -763,6 +763,58 @@ describe("service", function(){
       var match = URL_MATCH.exec('http://server/#?book=moby');
 
       expect(match[10]).toEqual('?book=moby');
+    });
+  });
+  
+  describe('$updateView', function(){
+    var scope, browser, evalCount;
+    
+    beforeEach(function(){
+      browser = new MockBrowser();
+      // Pretend that you are real Browser so that we see the delays
+      browser.isMock = false;
+      browser.defer = jasmine.createSpy('defer');
+
+      scope = angular.scope(null, null, {$browser:browser});
+      scope.$onEval(function(){ evalCount++; });
+      evalCount = 0;
+    });
+    
+    it('should eval root scope after a delay', function(){
+      scope.$updateView();
+      expect(evalCount).toEqual(0);
+      expect(browser.defer).toHaveBeenCalled();
+      expect(browser.defer.mostRecentCall.args[1]).toEqual(25);
+      browser.defer.mostRecentCall.args[0]();
+      expect(evalCount).toEqual(1);
+    });
+
+    it('should allow changing of delay time', function(){
+      var oldValue = angular.service('$updateView').delay;
+      angular.service('$updateView').delay = 50;
+      scope.$updateView();
+      expect(evalCount).toEqual(0);
+      expect(browser.defer).toHaveBeenCalled();
+      expect(browser.defer.mostRecentCall.args[1]).toEqual(50);
+      angular.service('$updateView').delay = oldValue;
+    });
+    
+    it('should ignore multiple requests for update', function(){
+      scope.$updateView();
+      scope.$updateView();
+      expect(evalCount).toEqual(0);
+      expect(browser.defer).toHaveBeenCalled();
+      expect(browser.defer.callCount).toEqual(1);
+      browser.defer.mostRecentCall.args[0]();
+      expect(evalCount).toEqual(1);
+    });
+    
+    it('should update immediatly in test/mock mode', function(){
+      scope = angular.scope();
+      scope.$onEval(function(){ evalCount++; });
+      expect(evalCount).toEqual(0);
+      scope.$updateView();
+      expect(evalCount).toEqual(1);
     });
   });
 });